refactor: add conditional to verify fk

+++ src/data/services/Location/DeleteLocation.service.js
@@ -1,8 +1,12 @@
-import { LocationRepository } from '../../database/repositories';
+import {
+  LocationRepository,
+  PolygonAreaRepository,
+} from '../../database/repositories';
 
 export class DeleteLocationService {
   async execute({ id_location }) {
     const repository = new LocationRepository();
+    const polygonAreaRepository = new PolygonAreaRepository();
 
     const verifyLocationExists = await repository.findLocationById({
       id_location,
@@ -12,6 +16,17 @@ export class DeleteLocationService {
       return {
         error: `Não há nenhuma Localização de Canteiro registrada com este ID -> ${id_location}.`,
       };
+
+    const verifyFk = await polygonAreaRepository.verifyLocation({
+      id_location,
+    });
+
+    if (verifyFk.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir a Localização de Canteiro pois existem Polígono de Área associados.',
+      };
+    }
 
     await repository.deleteLocation({
       id_location,
